Support calling BAM pack via 'python -m bam.pack'.
This allows us to run "setup.py bdist_wheel", put the wheel on the
PYTHONPATH, and run BAM pack without having to install it on a virtualenv.
The main purpose is to be able to ship this wheel with Blender. Updating
Blender's version of BAM thus requires only putting in a new wheel file,
rather than the old error-prone approach of copy-and-edit.